chain: move Memo into transaction
