--- zebra-chain/src/notes.rs
@@ -1,5 +1,0 @@
-//! Notes, note encryption, and nullifier types.
-
-#[cfg(test)]
-mod arbitrary;
-pub mod memo;
